+++ src/core/managers/PendingTransactionsStateManager/index.ts
@@ -1,2 +1,2 @@
-export * from './PendingTransactionsStateManagement';
+export * from './PendingTransactionsStateManager';
 export * from './types';