+++ src/core/providers/strategies/helpers/signTransactions/signTransactions.ts
@@ -28,13 +28,16 @@ import { getMultiEsdtTransferData } from './helpers/getMultiEsdtTransferData/get
 import { getScCall } from './helpers/getScCall';
 import { getTokenType } from './helpers/getTokenType';
 import { getUsdValue } from './helpers/getUsdValue';
+import { guardTransactions as getGuardedTransactions } from './helpers/guardTransactions/guardTransactions';
 
 export async function signTransactions({
   transactions = [],
-  handleSign
+  handleSign,
+  guardTransactions = getGuardedTransactions
 }: {
   transactions?: Transaction[];
   handleSign: IProvider['signTransactions'];
+  guardTransactions?: typeof getGuardedTransactions;
 }) {
   const address = getAddress();
   const network = networkSelector(getState());
@@ -232,14 +235,19 @@ export async function signTransactions({
 
           removeEvents();
 
-          if (signedTransactions.length == transactions.length) {
+          const areAllSigned = signedTransactions.length == transactions.length;
+
+          if (areAllSigned) {
+            const optionallyGuardedTransactions =
+              await guardTransactions(signedTransactions);
             signModalElement.remove();
-            resolve(signedTransactions);
-          } else {
-            currentTransactionIndex++;
-            manager.setNextUnsignedTxIndex(currentTransactionIndex);
-            signNextTransaction();
+
+            return resolve(optionallyGuardedTransactions);
           }
+
+          currentTransactionIndex++;
+          manager.setNextUnsignedTxIndex(currentTransactionIndex);
+          signNextTransaction();
         } catch (error) {
           reject('Error signing transactions: ' + error);
           signModalElement.remove();